portfolio
I am a dedicated IT student currently pursuing a Bachelor’s degree in Computer Science at RJ College. With a solid foundation in programming languages such as HTML ,CSS, Python, Java, and JavaScript. Looking ahead, I am eager to apply my skills in a professional setting, particularly in IT sector, and to contribute to innovative projects that drive technological advancement. I am committed to continuous learning and excited about the opportunities in the ever-evolving IT landscape